E2: per-site warning fixes -- 370 -> 308, five real bugs among them
The verified-each-site tranche: beam.cc's contradictory range guard
(< 0 && >= MAX, never rejected; the Assert above spells the intent),
popup.cc's bare-macro PF_ALLOW_DEAD_KEYS test (always true; respelled
to the flag test, matching fs2open's "unused even in retail" verdict),
freespace.cc's dangling jump_node_name read after its switch case,
medals.cc's self-overlapping sprintf, and ship.cc's dead
show_ship_subsys_count() debug stat hiding a Ships[objp->type]
indexing bug (deleted -- its high-water mark was never read).

The rest is inert respelling: char-subscript (int) casts, parentheses
documenting shipped precedence, braces pinning else bindings,
int-typed campaign file magics, the NULL-as-scalar family, and
is_training_mission gaining a proper declaration.  Delta fully
accounted: 62 removed (one medals statement carried two), 0 added.
Tests green, headless boot renders 73 frames.

+++ src/controlconfig/controlsconfig.cc
@@ -464,8 +464,8 @@ control_config_conflict_check()
                         if (a == b) {
                             Conflicts[i].key = j;
                             Conflicts[j].key = i;
-                            Conflicts_tabs[Control_config[i].tab] = 1;
-                            Conflicts_tabs[Control_config[j].tab] = 1;
+                            Conflicts_tabs[(int)Control_config[i].tab] = 1;
+                            Conflicts_tabs[(int)Control_config[j].tab] = 1;
                         }
 
                         /*          if ((a >= 0) && (a & KEY_SHIFTED) && (shift >= 0)) {
@@ -501,8 +501,8 @@ control_config_conflict_check()
                         (Control_config[i].joy_id == Control_config[j].joy_id)) {
                         Conflicts[i].joy = j;
                         Conflicts[j].joy = i;
-                        Conflicts_tabs[Control_config[i].tab] = 1;
-                        Conflicts_tabs[Control_config[j].tab] = 1;
+                        Conflicts_tabs[(int)Control_config[i].tab] = 1;
+                        Conflicts_tabs[(int)Control_config[j].tab] = 1;
                     }
                 }
             }
